bugfix(extract): consistently treat .d.ts as a separate extension
This fixes the following bug:

Steps taken
-----------
- take 2 .d.ts modules where the one includes the other (one.d.ts => two.d.ts)
- set the options.enhancedResolveOptions.extensions to [".d.ts"] (so DONOT include ".ts")
- cruise

Expected
--------
two.d.ts is a dependency of one.d.ts
two.d.ts' 'followable' attribute equals true

Found
-----
two.d.ts is a dependency of one.d.ts (OK)
two.d.ts' 'followable' attribute equals false (KO)

Background
----------

Before this bugfix the code determining whether an extension was followable looked
at the extensions array and compared them with the path.extname (except for .coffee.md
which is simmilarly weird). As ".ts" !== ".d.ts" followable

A ripple effect was that .d.ts dependencies were 'completed' into the module graph as
modules without dependencies of themselves. Often they then popped up a second time
with a bunch of dependencies of their own, so there were _two_ of them in the resulting
module graph. Which is confusing when you do a modules.find and the first,
dependency-less .d.ts shows up only.

Severity
--------

The amount of repo's that (1) define an array of enhancedResolveOptions.extensions and
(2) _only_ have .d.ts in there and no .ts is likely not too big (dependency-crduiser
itself was one of them, though) - but it's a breaking change nonetheless, as e.g.
the cyclic dependency-algorithm now will find more.
